Ported all effects/components/modifiers/attributes to assets
