+++ assets/moves/components/SelfCast.py
@@ -1,0 +1,6 @@
+from engine.game.move.component import Component
+
+class SelfCast(Component):
+    """Targeting scheme for self only"""
+    def get_targets(self, selected, caster, players, monsters):
+        return [caster]
