+++ engine/game/move/built_modifiers.py
@@ -1,40 +1,3 @@
 """Compilation of the modifiers to be applied to move values"""
 
-from engine.game.move.modifier import Modifier
-
-class ScaleStat(Modifier):
-    """Additively scales value by a percentage of a stat"""
-
-    def __init__(self, scaling, stype):
-        self.scaling = scaling
-        self.stype = stype
-
-    def modify(self, value, target, caster):
-        return value + self.scaling * caster.get_stat(self.stype)
-
-class ScaleLevel(Modifier):
-    """Multiplicatively scales value by a percentage of level"""
-
-    def __init__(self, scaling):
-        self.scaling = scaling
-
-    def modify(self, value, target, caster):
-        return value * (self.scaling * caster.get_level())
-
-class ScaleLevelAdd(Modifier):
-    """Additively scales value by a percentage"""
-
-    def __init__(self, scaling):
-        self.scaling = scaling
-
-    def modify(self, value, target, caster):
-        return value + self.scaling * caster.get_level()
-
-class ScaleCritDamage(Modifier):
-    """Multiplicatively scales value by a percentage"""
-
-    def __init__(self, scaling):
-        self.scaling = scaling
-
-    def modify(self, value, target, caster):
-        return value * self.scaling
+from assets.moves.modifiers import *
